new CommonTools/Egamma package
