+++ CommonTools/Egamma/src/EffectiveAreas.cc
@@ -1,4 +1,4 @@
-#include "CommonTools/Utils/interface/EffectiveAreas.h"
+#include "CommonTools/Egamma/interface/EffectiveAreas.h"
 #include "FWCore/Utilities/interface/Exception.h"
 
 #include <cmath>
