+++ Configuration/Skimming/BuildFile.xml
@@ -1,3 +1,4 @@
+<use name="CommonTools/Egamma"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
@@ -12,7 +13,6 @@
 <use name="RecoJets/JetAlgorithms"/>
 <use name="HLTrigger/HLTcore"/>
 <use name="CommonTools/UtilAlgos"/>
-<use name="RecoEgamma/EgammaTools"/>
 <use name="PhysicsTools/UtilAlgos"/>
 <use name="DataFormats/Candidate"/>
 <use name="DataFormats/Luminosity"/>
