+++ Configuration/Skimming/interface/LeptonSkimming.h
@@ -36,7 +36,7 @@
 #include "DataFormats/VertexReco/interface/Vertex.h"
 #include "DataFormats/EgammaCandidates/interface/ConversionFwd.h"
 #include "DataFormats/EgammaCandidates/interface/Conversion.h"
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 #include "DataFormats/Common/interface/TriggerResults.h"
 #include "FWCore/Common/interface/TriggerNames.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
@@ -52,7 +52,6 @@
 #include "DataFormats/RecoCandidate/interface/RecoChargedCandidateIsolation.h"
 #include "DataFormats/Common/interface/AssociationMap.h"
 #include "HLTrigger/HLTcore/interface/HLTFilter.h"
-//#include "HLTrigger/Egamma/plugins/HLTGenericFilter.h"
 #include "DataFormats/HLTReco/interface/TriggerFilterObjectWithRefs.h"
 #include "DataFormats/TrackReco/interface/Track.h"
 #include "DataFormats/MuonReco/interface/Muon.h"
