+++ DQMOffline/L1Trigger/interface/L1TStage2CaloLayer2Offline.h
@@ -40,7 +40,7 @@
 #include "DataFormats/METReco/interface/CaloMETCollection.h"
 
 // Conversions
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 
 // Trigger
 #include "DataFormats/Common/interface/TriggerResults.h"
