+++ DQMOffline/Trigger/plugins/BuildFile.xml
@@ -9,7 +9,7 @@
 <use name="DataFormats/TauReco"/>
 <use name="DataFormats/METReco"/>
 <use name="DataFormats/JetReco"/>
-<use name="RecoEgamma/EgammaTools"/>
+<use name="CommonTools/Egamma"/>
 <use name="RecoVertex/KalmanVertexFit"/>
 <use name="HLTrigger/HLTcore"/>
 <use name="DQMServices/Core"/>
