+++ DQMServices/Examples/BuildFile.xml
@@ -1,10 +1,10 @@
+<use name="CommonTools/Egamma"/>
 <use name="DQMServices/Core"/>
 <use name="FWCore/Framework"/>
 <use name="DataFormats/EgammaCandidates"/>
 <use name="DataFormats/JetReco"/>
 <use name="DataFormats/METReco"/>
 <use name="DataFormats/VertexReco"/>
-<use name="RecoEgamma/EgammaTools"/>
 <export>
   <lib name="1"/>
 </export>
