+++ EgammaAnalysis/ElectronTools/BuildFile.xml
@@ -1,3 +1,4 @@
+<use name="CommonTools/Egamma"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
@@ -7,7 +8,6 @@
 <use name="CondCore/DBOutputService"/>
 <use name="CondFormats/EgammaObjects"/>
 <use name="CondFormats/PhysicsToolsObjects"/>
-<use name="RecoEgamma/EgammaTools"/>
 <use name="RecoEcal/EgammaCoreTools"/>
 <use name="DataFormats/Candidate"/>
 <use name="DataFormats/Common"/>
@@ -18,6 +18,7 @@
 <use name="TrackingTools/IPTools"/>
 <use name="PhysicsTools/UtilAlgos"/>
 <use name="PhysicsTools/SelectorUtils"/>
+<use name="RecoEgamma/EgammaTools"/>
 <use name="clhep"/>
 <use name="root"/>
 <use name="rootcore"/>
