+++ ElectroWeakAnalysis/WENu/BuildFile.xml
@@ -3,7 +3,6 @@
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/Common"/>
-<use name="RecoEgamma/EgammaTools"/>
 <use name="RecoLocalCalo/EcalRecAlgos"/>
 <use name="root"/>
 <use name="rootmath"/>
