+++ ElectroWeakAnalysis/ZEE/BuildFile.xml
@@ -6,12 +6,13 @@
 <use name="DataFormats/Scalers"/>
 <use name="Geometry/CaloGeometry"/>
 <use name="Geometry/CaloTopology"/>
+<use name="CommonTools/Egamma"/>
 <use name="CommonTools/UtilAlgos"/>
 <use name="L1Trigger/GlobalTriggerAnalyzer"/>
 <use name="MagneticField/Records"/>
 <use name="MagneticField/Engine"/>
-<use name="RecoEgamma/EgammaTools"/>
 <use name="RecoEcal/EgammaCoreTools"/>
+<use name="RecoEgamma/EgammaTools"/>
 <use name="RecoLocalCalo/EcalRecAlgos"/>
 <use name="HLTrigger/HLTcore"/>
 <use name="root"/>
