+++ HLTriggerOffline/SUSYBSM/src/SUSY_HLT_SingleLepton.cc
@@ -15,7 +15,7 @@
 
 #include "DataFormats/EgammaCandidates/interface/Conversion.h"
 #include "DataFormats/EgammaCandidates/interface/ConversionFwd.h"
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 
 namespace {
   typedef dqm::reco::DQMStore DQMStore;
