+++ PhysicsTools/NanoAOD/plugins/BuildFile.xml
@@ -5,6 +5,7 @@
 <use name="FWCore/Utilities"/>
 <use name="DataFormats/Candidate"/>
 <use name="DataFormats/PatCandidates"/>
+<use name="CommonTools/Egamma"/>
 <use name="CommonTools/MVAUtils"/>
 <use name="PhysicsTools/JetMCUtils"/>
 <use name="PhysicsTools/PatAlgos"/>
