+++ PhysicsTools/PatAlgos/plugins/BuildFile.xml
@@ -22,17 +22,18 @@
   <use name="CondFormats/HcalObjects"/>
   <use name="CondFormats/DataRecord"/>
   <use name="CommonTools/CandAlgos"/>
+  <use name="CommonTools/Egamma"/>
   <use name="CommonTools/MVAUtils"/>
   <use name="JetMETCorrections/Objects"/>
   <use name="JetMETCorrections/JetCorrector"/>
   <use name="TrackingTools/TransientTrack"/>
+  <use name="RecoEgamma/EgammaTools"/>
   <use name="SimDataFormats/Track"/>
   <use name="SimDataFormats/Vertex"/>
   <use name="SimDataFormats/JetMatching"/>
   <use name="SimDataFormats/PileupSummaryInfo"/>
   <use name="SimGeneral/HepPDTRecord"/>
   <use name="RecoMET/METAlgorithms"/>
-  <use name="RecoEgamma/EgammaTools"/>
   <use name="RecoTracker/DeDx"/>
   <use name="TrackingTools/IPTools"/>
   <use name="root"/>
