+++ PhysicsTools/PatAlgos/plugins/PATPhotonProducer.cc
@@ -16,7 +16,7 @@
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
 
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 #include "RecoEgamma/EgammaTools/interface/EcalRegressionData.h"
 #include "RecoEcal/EgammaCoreTools/interface/EcalClusterTools.h"
