+++ RecoEgamma/EgammaPhotonProducers/src/ReducedEGProducer.cc
@@ -38,7 +38,7 @@
 
 #include "RecoEgamma/EgammaIsolationAlgos/interface/EgammaHadTower.h"
 
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 
 ReducedEGProducer::ReducedEGProducer(const edm::ParameterSet& config)
     : photonT_(consumes<reco::PhotonCollection>(config.getParameter<edm::InputTag>("photons"))),
