+++ RecoEgamma/EgammaTools/BuildFile.xml
@@ -1,3 +1,5 @@
+<use name="CommonTools/Egamma"/>
+<use name="CommonTools/Utils"/>
 <use name="FWCore/Framework"/>
 <use name="DataFormats/HcalRecHit"/>
 <use name="DataFormats/EgammaReco"/>
@@ -11,7 +13,6 @@
 <use name="CondFormats/EgammaObjects"/>
 <use name="RecoEcal/EgammaCoreTools"/>
 <use name="DataFormats/ParticleFlowReco"/>
-<use name="CommonTools/Utils"/>
 <export>
   <lib name="1"/>
 </export>
