+++ RecoEgamma/EgammaTools/interface/EffectiveAreas.h
@@ -1,0 +1,1 @@
+#warning "Obsolete header, include CommonTools/Egamma/interface/EffectiveAreas.h instead"
