+++ RecoEgamma/EgammaTools/plugins/BuildFile.xml
@@ -2,6 +2,7 @@
   <flags EDM_PLUGIN="1"/>
   <use name="FWCore/Framework"/>
   <use name="CommonTools/CandAlgos"/>
+  <use name="CommonTools/Egamma"/>
   <use name="CondCore/DBOutputService"/>
   <use name="CondFormats/DataRecord"/>
   <use name="CondFormats/EgammaObjects"/>
