+++ RecoEgamma/ElectronIdentification/BuildFile.xml
@@ -1,5 +1,6 @@
 <use name="CondFormats/EgammaObjects"/>
 <use name="DataFormats/TrackReco"/>
+<use name="CommonTools/Egamma"/>
 <use name="CommonTools/MVAUtils"/>
 <use name="RecoEgamma/EgammaTools"/>
 <use name="PhysicsTools/SelectorUtils"/>
