+++ RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleCalPFClusterIsoCut.cc
@@ -1,6 +1,6 @@
 #include "PhysicsTools/SelectorUtils/interface/CutApplicatorWithEventContentBase.h"
 #include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
-#include "CommonTools/Utils/interface/EffectiveAreas.h"
+#include "CommonTools/Egamma/interface/EffectiveAreas.h"
 
 class GsfEleCalPFClusterIsoCut : public CutApplicatorWithEventContentBase {
 public:
