+++ RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleConversionVetoCut.cc
@@ -2,7 +2,7 @@
 #include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
 #include "DataFormats/EgammaCandidates/interface/ConversionFwd.h"
 #include "DataFormats/EgammaCandidates/interface/Conversion.h"
-#include "RecoEgamma/EgammaTools/interface/ConversionTools.h"
+#include "CommonTools/Egamma/interface/ConversionTools.h"
 
 class GsfEleConversionVetoCut : public CutApplicatorWithEventContentBase {
 public:
