+++ RecoEgamma/PhotonIdentification/BuildFile.xml
@@ -1,3 +1,4 @@
+<use name="CommonTools/Egamma"/>
 <use name="FWCore/Framework"/>
 <use name="DataFormats/HcalRecHit"/>
 <use name="DataFormats/EgammaReco"/>
